refactor: Move evaluation to module
